Add Pry and make some corrections. Specify things in Readme.
Add Pry to replace IRB as the console.
Some names in files were still according to how the app was generated, rather than the current app name.
This is corrected here.
The Readme is now more complete, including some steps on setup, i.e. what needs to be changed in which file
when starting with a new project is specified.